deps/v8/test/inspector/runtime/console-messages-limits-expected.txt

Checks that console message storage doesn't exceed limits

Running test: testMaxConsoleMessagesCount
Messages reported: 1000

Running test: testMaxConsoleMessagesV8Size
Messages reported: 3
